+++ src/components/App/App.tsx
@@ -1,9 +1,6 @@
 import * as React from 'react';
+import { Timer } from '../Timer/Timer';
 
-export interface HelloProps { compiler: string; framework: string }
-
-export class App extends React.Component<HelloProps, {}> {
-    render(): JSX.Element {
-        return <h1>Hello from {this.props.compiler} and {this.props.framework}!</h1>;
-    }
+export class App extends React.Component {
+    render(): JSX.Element { return <Timer />; }
 }