+++ src/components/Timer/Timer.tsx
@@ -1,3 +1,3 @@
 import * as React from 'react';
 
-export const Timer = (): JSX.Element => <div></div>;
+export const Timer = (): JSX.Element => <div>Timer</div>;